+++ src/scriptlets/set-cookie-reload.js
@@ -29,13 +29,19 @@ import {
  * - `value` — required, cookie value; possible values:
  *     - number `>= 0 && <= 15`
  *     - one of the predefined constants in any case variation:
- *         - `true`
- *         - `false`
+ *         - `true` / `t`
+ *         - `false` / `f`
  *         - `yes` / `y`
  *         - `no` / `n`
  *         - `ok`
- *         - `accept`/ `reject`
- *         - `allow` / `deny`
+ *         - `on` / `off`
+ *         - `accept`/ `accepted` / `notaccepted`
+ *         - `reject` / `rejected`
+ *         - `allow` / `allowed`
+ *         - `disallow` / `deny`
+ *         - `enable` / `enabled`
+ *         - `disable` / `disabled`
+ *         - `necessary` / `required`
  * - `path` — optional, cookie path, defaults to `/`; possible values:
  *     - `/` — root path
  *     - `none` — to set no path at all