+++ tests/scriptlets/set-cookie.test.js
@@ -38,163 +38,37 @@ test('Checking if alias name works', (assert) => {
     assert.strictEqual(codeByAdgParams, codeByUboParams, 'ubo name - ok');
 });
 
-test('Set cookie with valid value', (assert) => {
-    let cName = '__test-cookie_OK';
-    let cValue = 'OK';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
+const cookies = [
+    ['__test-cookie_OK', 'OK'],
+    ['__test-cookie_true', 'true'],
+    ['__test-cookie_true', 't'],
+    ['__test-cookie_false', 'false'],
+    ['__test-cookie_false', 'f'],
+    ['__test-cookie_no', 'no'],
+    ['__test-cookie_n', 'n'],
+    ['__test-cookie_Accept', 'Accept'],
+    ['__test-cookie_Reject', 'Reject'],
+    ['__test-cookie_allow', 'allow'],
+    ['__test-cookie_deny', 'deny'],
+    ['__test-cookie_dENy', 'dENy'],
+    ['__test-cookie_0', '0'],
+    ['__test-cookie_1', '1'],
+    ['__test-cookie_on', 'on'],
+    ['__test-cookie_off', 'off'],
+    ['__test-cookie_accepted', 'accepted'],
+    ['__test-cookie_notaccepted', 'notaccepted'],
+    ['__test-cookie_rejected', 'rejected'],
+    ['__test-cookie_allowed', 'allowed'],
+    ['__test-cookie_disallow', 'disallow'],
+    ['__test-cookie_enable', 'enable'],
+    ['__test-cookie_enabled', 'enabled'],
+    ['__test-cookie_disable', 'disable'],
+    ['__test-cookie_disabled', 'disabled'],
+    ['__test-cookie_necessary', 'necessary'],
+    ['__test-cookie_required', 'required'],
+];
 
-    cName = '__test-cookie_true';
-    cValue = 'true';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_false';
-    cValue = 'false';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_no';
-    cValue = 'no';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_n';
-    cValue = 'n';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_Accept';
-    cValue = 'Accept';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_Reject';
-    cValue = 'Reject';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_allow';
-    cValue = 'allow';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_deny';
-    cValue = 'deny';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_dENy';
-    cValue = 'dENy';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_0';
-    cValue = '0';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_1';
-    cValue = '1';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_on';
-    cValue = 'on';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_off';
-    cValue = 'off';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_accepted';
-    cValue = 'accepted';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_notaccepted';
-    cValue = 'notaccepted';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_rejected';
-    cValue = 'rejected';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_allowed';
-    cValue = 'allowed';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_disallow';
-    cValue = 'disallow';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_enable';
-    cValue = 'enable';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_enabled';
-    cValue = 'enabled';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_disable';
-    cValue = 'disable';
-    runScriptlet(name, [cName, cValue]);
-    assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
-    assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');
-    clearCookie(cName);
-
-    cName = '__test-cookie_disabled';
-    cValue = 'disabled';
+test.each('Set cookie with valid value', cookies, (assert, [cName, cValue]) => {
     runScriptlet(name, [cName, cValue]);
     assert.strictEqual(window.hit, 'FIRED', 'Hit was fired');
     assert.strictEqual(document.cookie.includes(cName) && document.cookie.includes(cValue), true, 'Cookie is set');